use C++20 add copyright to NewMoon files
move some headery stuff to header, add some todo notes
