+++ src/CMakeLists.txt
@@ -473,14 +473,14 @@ IF(GENERATE_STELMAINLIB)
      ADD_EXECUTABLE(stellarium WIN32 ${stellarium_exe_SRCS})
      #stelMain should be after the plug-ins, otherwise the build crashes
      TARGET_LINK_LIBRARIES(stellarium stelMain ${winMMLib})
-     ADD_EXECUTABLE(newmoon WIN32 newmoon.cpp)
+     ADD_EXECUTABLE(newmoon WIN32 NewMoon.cpp NewMoon.hpp)
      TARGET_LINK_LIBRARIES(newmoon stelMain ${winMMLib})
 ELSE()
      ADD_LIBRARY(stelMain STATIC ${stellarium_lib_SRCS} ${stellarium_RES_CXX})
      TARGET_LINK_LIBRARIES(stelMain ${STELMAIN_DEPS})
      ADD_EXECUTABLE(stellarium WIN32 ${stellarium_exe_SRCS})
      TARGET_LINK_LIBRARIES(stellarium ${winMMLib} ${STELMAIN_DEPS} stelMain)
-     ADD_EXECUTABLE(newmoon WIN32 newmoon.cpp)
+     ADD_EXECUTABLE(newmoon WIN32 NewMoon.cpp NewMoon.hpp)
      TARGET_LINK_LIBRARIES(newmoon ${winMMLib} ${STELMAIN_DEPS} stelMain)
 
      IF(APPLE)
